+++ src/Components/Main/MainRoutes.js
@@ -6,16 +6,16 @@ import { UsersPageContainer } from './Pages/UsersPage/UsersPageContainer';
 import { MessagesPage } from './Pages/MessagesPage/MessagesPage';
 import { useAuthRedirect } from '../../hooks/useRedirect';
 import { LoginPageContainer } from './Pages/LoginPage/LoginPageContainer';
-import { ProfilePageContainer } from './Pages/ProfilePage/ProfilePageContainer';
+import { ProfilePage } from './Pages/ProfilePage/ProfilePage';
 
 export const MainRoutes = () => {
-  const authProfilePage = useAuthRedirect(<ProfilePageContainer /> );
+  const authProfilePage = useAuthRedirect(<ProfilePage /> );
   const authMessagesPage = useAuthRedirect(<MessagesPage /> );
 
   return (
     <Routes>
       <Route path="/profile" element={authProfilePage} />
-      <Route path="/profile/:id?" element={<ProfilePageContainer />} />
+      <Route path="/profile/:id?" element={<ProfilePage />} />
       <Route path="/messages/*" element={authMessagesPage} />
       <Route path="/users" element={<UsersPageContainer />} />
       <Route path="/news" element={<NewsPage />} />
