+++ src/Components/Main/Pages/ProfilePage/ProfilePage.js
@@ -1,15 +1,12 @@
 import classes from './ProfilePage.module.scss';
 import { PostsContainer } from './Posts/PostsContainer';
 import { ProfileInfoContainer } from './ProfileInfo/ProfileInfoContainer';
-import { Preloader } from '../../../Common/Preloader/Preloader';
 
-export const ProfilePage = ({isFetchingProfile}) => {
+export const ProfilePage = () => {
   return (
-    isFetchingProfile ?
-      <Preloader /> :
-      <div className={classes.profilePageBlock}>
-        <ProfileInfoContainer />
-        <PostsContainer />
-      </div>
+    <div className={classes.profilePageBlock}>
+      <ProfileInfoContainer />
+      <PostsContainer />
+    </div>
   );
 };
