--- src/Components/Main/Pages/ProfilePage/ProfilePageContainer.js
@@ -1,34 +1,0 @@
-import { ProfilePage } from './ProfilePage';
-import { useDispatch, useSelector } from 'react-redux';
-import { useEffect } from 'react';
-import { getProfile, getStatus } from '../../../../redux/thunks/profile';
-import { useNavigate, useParams } from 'react-router-dom';
-import { isFetchingProfileSelector } from '../../../../redux/selectors/loading';
-import { idSelector } from '../../../../redux/selectors/auth';
-
-export const ProfilePageContainer = () => {
-  const isFetchingProfile = useSelector(isFetchingProfileSelector);
-  const authorizedUserId = useSelector(idSelector);
-
-  const dispatch = useDispatch();
-
-  let {id: paramId} = useParams();
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    if (!paramId) {
-      paramId = authorizedUserId;
-      if (!paramId) {
-        navigate('/login');
-      }
-    }
-    if (paramId) {
-      dispatch(getProfile(paramId));
-      dispatch(getStatus(paramId));
-    }
-  }, [paramId]);
-
-  return (
-    <ProfilePage isFetchingProfile={isFetchingProfile} />
-  );
-};
